Find the longest match
Now I can more freely order patterns, although it still matches only the
first matcher within a pattern. That's much less of a potential issue as
individual library authors could sort that out (even if it's a pain)
whereas you can't control the order of patterns because libraries should
be able to be included in any order.